In progress

+++ tests/fixtures/database.py
@@ -2,6 +2,7 @@ import logging
 from collections.abc import Iterator
 
 import pytest
+from _pytest.fixtures import FixtureRequest
 from sqlalchemy import Engine, create_engine
 from sqlalchemy.orm import Session
 
@@ -12,24 +13,25 @@ logger = logging.getLogger(__name__)
 
 
 @pytest.fixture(scope="session")
-def engine(settings: Settings) -> Iterator[Engine]:
-    logger.debug(f"Creating engine {settings.SQLALCHEMY_DATABASE_URI}")
+def engine(settings: Settings) -> Engine:
+    logger.info(f"Create engine {settings.SQLALCHEMY_DATABASE_URI}")
     engine = create_engine(str(settings.SQLALCHEMY_DATABASE_URI))
-    logger.debug("Creating database")
-    Base.metadata.create_all(engine)
-    yield engine
-    logger.debug("Deleting database")
+    logger.info("Drop all tables")
     Base.metadata.drop_all(engine)
+    logger.info("Create all tables")
+    Base.metadata.create_all(engine)
+    return engine
 
 
 @pytest.fixture
-def session(engine: Engine) -> Iterator[Session]:
-    with Session(engine) as session:
+def session(request: FixtureRequest, engine: Engine) -> Iterator[Session]:
+    with Session(engine, expire_on_commit=False) as session:
         logger.debug("Creating session")
         yield session
 
-        session.rollback()
-        logger.debug("Deleting database tables")
-        for table in reversed(Base.metadata.sorted_tables):
-            session.execute(table.delete())
-        session.commit()
+        if not request.config.getoption("--db"):
+            session.rollback()
+            logger.debug("Deleting database tables")
+            for table in reversed(Base.metadata.sorted_tables):
+                session.execute(table.delete())
+            session.commit()
